[refactor] store hook에서 사용하는 공통 함수 분리
