+++ FE/src/store/AccountBook/accountBookInfoHook.ts
@@ -1,22 +1,6 @@
-import React from 'react';
-import { useObserver } from 'mobx-react';
-
 import { transactionInfoContext } from './accountBookDataContext';
 import { TStore } from './accountBookData';
-
-export const useStoreData = <Selection, ContextData, Store>(
-  context: React.Context<ContextData>,
-  storeSelector: (ContextData: ContextData) => Store,
-  dataSelector: (store: Store) => Selection,
-) => {
-  const value = React.useContext(context);
-  if (!value) throw new Error();
-
-  const store = storeSelector(value);
-  return useObserver(() => {
-    return dataSelector(store);
-  });
-};
+import { useStoreData } from '../useStoreData';
 
 export const useTransactionData = <Selection>(
   dataSelector: (store: TStore) => Selection,
