+++ FE/src/store/DateInfo/dateInfoHook.ts
@@ -1,21 +1,6 @@
-import React from 'react';
-import { useObserver } from 'mobx-react';
 import { dateInfoContext } from './dateInfoContext';
 import { TStore } from './dateInfoStore';
-
-export const useStoreData = <Selection, ContextData, Store>(
-  context: React.Context<ContextData>,
-  storeSelector: (ContextData: ContextData) => Store,
-  dataSelector: (store: Store) => Selection,
-) => {
-  const value = React.useContext(context);
-  if (!value) throw new Error();
-
-  const store = storeSelector(value);
-  return useObserver(() => {
-    return dataSelector(store);
-  });
-};
+import { useStoreData } from '../useStoreData';
 
 export const useRootData = <Selection>(
   dataSelector: (store: TStore) => Selection,
